Relocate conversion utils to shared/modules/conversion.utils.js
