+++ app/scripts/controllers/detect-tokens.test.js
@@ -9,9 +9,9 @@ import {
   TokensController,
   AssetsContractController,
 } from '@metamask/assets-controllers';
+import { convertHexToDecimal } from '@metamask/controller-utils';
 import { NETWORK_TYPES } from '../../../shared/constants/network';
 import { toChecksumHexAddress } from '../../../shared/modules/hexstring-utils';
-import { hexToDecimal } from '../../../shared/lib/metamask-controller-utils';
 import DetectTokensController from './detect-tokens';
 import NetworkController, { NETWORK_EVENTS } from './network';
 import PreferencesController from './preferences';
@@ -88,7 +88,7 @@ describe('DetectTokensController', function () {
             ...networkState,
             providerConfig: {
               ...networkState.provider,
-              chainId: hexToDecimal(networkState.provider.chainId),
+              chainId: convertHexToDecimal(networkState.provider.chainId),
             },
           };
           return cb(modifiedNetworkState);
