+++ app/scripts/controllers/transactions/tx-gas-utils.test.js
@@ -1,7 +1,8 @@
 import { strict as assert } from 'assert';
 import { TransactionFactory } from '@ethereumjs/tx';
 import Common from '@ethereumjs/common';
-import { hexToBn, bnToHex } from '../../lib/util';
+import { hexToBn } from '../../lib/util';
+import { bnToHex } from '../../../../shared/modules/conversion.utils';
 import TxUtils from './tx-gas-utils';
 
 describe('txUtils', function () {
