+++ ui/components/app/advanced-gas-inputs/advanced-gas-inputs.container.js
@@ -5,7 +5,7 @@ import { MIN_GAS_LIMIT_DEC } from '../../../pages/send/send.constants';
 import {
   decimalToHex,
   hexWEIToDecGWEI,
-} from '../../../../shared/lib/transactions-controller-utils';
+} from '../../../../shared/modules/conversion.utils';
 import AdvancedGasInputs from './advanced-gas-inputs.component';
 
 function convertGasPriceForInputs(gasPriceInHexWEI) {
