+++ ui/components/app/cancel-speedup-popover/cancel-speedup-popover.test.js
@@ -13,7 +13,7 @@ import { GasFeeContextProvider } from '../../../contexts/gasFee';
 import configureStore from '../../../store/store';
 import { decGWEIToHexWEI } from '../../../helpers/utils/conversions.util';
 import InfoTooltip from '../../ui/info-tooltip';
-import { hexWEIToDecETH } from '../../../../shared/lib/transactions-controller-utils';
+import { hexWEIToDecETH } from '../../../../shared/modules/conversion.utils';
 import CancelSpeedupPopover from './cancel-speedup-popover';
 
 const MAXFEEPERGAS_ABOVE_MOCK_MEDIUM_HEX = '0x174876e800';
