+++ ui/components/app/transaction-breakdown/transaction-breakdown.container.js
@@ -5,10 +5,12 @@ import {
 } from '../../../selectors';
 import { getNativeCurrency } from '../../../ducks/metamask/metamask';
 import { getHexGasTotal } from '../../../helpers/utils/confirm-tx.util';
-import { subtractHexes } from '../../../helpers/utils/conversions.util';
-import { sumHexes } from '../../../helpers/utils/transactions.util';
 import { isEIP1559Transaction } from '../../../../shared/modules/transaction.utils';
 
+import {
+  subtractHexes,
+  sumHexes,
+} from '../../../../shared/modules/conversion.utils';
 import TransactionBreakdown from './transaction-breakdown.component';
 
 const mapStateToProps = (state, ownProps) => {
