+++ ui/components/ui/hex-to-decimal/hex-to-decimal.component.js
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import { hexToDecimal } from '../../../../shared/lib/metamask-controller-utils';
+import { hexToDecimal } from '../../../../shared/modules/conversion.utils';
 
 export default class HexToDecimal extends PureComponent {
   static propTypes = {
