+++ ui/ducks/swaps/swaps.js
@@ -51,12 +51,14 @@ import {
   parseSmartTransactionsError,
   stxErrorTypes,
 } from '../../pages/swaps/swaps.util';
+import { decGWEIToHexWEI } from '../../helpers/utils/conversions.util';
 import {
-  getValueFromWeiHex,
-  decGWEIToHexWEI,
   addHexes,
-} from '../../helpers/utils/conversions.util';
-import { conversionLessThan } from '../../../shared/modules/conversion.utils';
+  conversionLessThan,
+  decimalToHex,
+  getValueFromWeiHex,
+  hexWEIToDecGWEI,
+} from '../../../shared/modules/conversion.utils';
 import {
   getSelectedAccount,
   getTokenExchangeRates,
@@ -88,8 +90,6 @@ import { ORIGIN_METAMASK } from '../../../shared/constants/app';
 import {
   calcGasTotal,
   calcTokenAmount,
-  decimalToHex,
-  hexWEIToDecGWEI,
 } from '../../../shared/lib/transactions-controller-utils';
 
 export const GAS_PRICES_LOADING_STATES = {
