+++ ui/hooks/gasFeeInput/useGasEstimates.js
@@ -16,7 +16,7 @@ import { isLegacyTransaction } from '../../helpers/utils/transactions.util';
 
 import { useCurrencyDisplay } from '../useCurrencyDisplay';
 import { useUserPreferencedCurrency } from '../useUserPreferencedCurrency';
-import { decimalToHex } from '../../../shared/lib/transactions-controller-utils';
+import { decimalToHex } from '../../../shared/modules/conversion.utils';
 
 /**
  * @typedef {object} GasEstimatesReturnType
