+++ ui/selectors/confirm-transaction.js
@@ -2,12 +2,10 @@ import { createSelector } from 'reselect';
 import txHelper from '../helpers/utils/tx-helper';
 import {
   roundExponential,
-  getValueFromWeiHex,
   getTransactionFee,
   addFiat,
   addEth,
 } from '../helpers/utils/confirm-tx.util';
-import { sumHexes } from '../helpers/utils/transactions.util';
 import { transactionMatchesNetwork } from '../../shared/modules/transaction.utils';
 import {
   getGasEstimateType,
@@ -26,6 +24,10 @@ import {
 } from '../../shared/modules/gas.utils';
 import { isEqualCaseInsensitive } from '../../shared/modules/string-utils';
 import { calcTokenAmount } from '../../shared/lib/transactions-controller-utils';
+import {
+  getValueFromWeiHex,
+  sumHexes,
+} from '../../shared/modules/conversion.utils';
 import { getAveragePriceEstimateInHexWEI } from './custom-gas';
 import { getCurrentChainId, deprecatedGetCurrentNetworkId } from './selectors';
 import { checkNetworkAndAccountSupports1559 } from '.';
